requirements.txt
opencv-python
python-dotenv

